#15: restructure codebase for src and test config
