+++ server.js
@@ -3,8 +3,8 @@ import express, { urlencoded } from "express";
 import http from "http";
 import dotenv from "dotenv";
 import { Server } from "socket.io";
-import router from "./routes/router.js";
-import socket from "./routes/socket.router.js";
+import router from "./src/routes/router.js";
+import socket from "./src/routes/socket.router.js";
 
 process.env.NODE_ENV !== "production" && dotenv.config();
